update types and arrange them into types folder, implement exhibition type
